+++ libs/sdk/src/lib/helpers/index.ts
@@ -1,6 +1,3 @@
 export * from './get-solana-rpc-endpoint'
 export * from './kin-to-quarks'
 export * from './parse-kinetic-sdk-endpoint'
-export * from './serialize-create-account-transaction'
-export * from './serialize-make-transfer-batch-transactions'
-export * from './serialize-make-transfer-transaction'
