+++ libs/solana/src/lib/helpers/index.ts
@@ -6,3 +6,6 @@ export * from './parse-and-sign-transaction'
 export * from './parse-endpoint'
 export * from './transaction-error'
 export * from './parse-transaction-simulation'
+export * from './serialize-create-account-transaction'
+export * from './serialize-make-transfer-batch-transactions'
+export * from './serialize-make-transfer-transaction'
