+++ libs/solana/src/lib/helpers/serialize-create-account-transaction.ts
@@ -1,9 +1,10 @@
 import { Keypair } from '@kin-kinetic/keypair'
-import { getPublicKey, PublicKeyString } from '@kin-kinetic/solana'
 import { createAssociatedTokenAccountInstruction, getAssociatedTokenAddress } from '@solana/spl-token'
 import { Transaction, TransactionInstruction } from '@solana/web3.js'
 import { TransactionType } from '@kin-tools/kin-memo'
 import { generateKinMemoInstruction } from '@kin-tools/kin-transaction'
+import { PublicKeyString } from '../interfaces'
+import { getPublicKey } from './get-public-key'
 
 export async function serializeCreateAccountTransaction({
   appIndex,
